requirements.txt
Flask
nltk
google-generativeai
python-dotenv
gunicorn